feat: add two-factor authentication settings
This commit adds a two-factor authentication (2FA) setting to the profile controls in
frontend-users.xml. It also includes the corresponding localization files in de and en to
locale.xml for these settings and messages. Additionally, some CSS properties for the 2FA elements
have been set for proper display of the 2FA settings.